requirements.txt
PyYAML==6.0.2
tabulate==0.9.0
tqdm==4.67.1

# install torch manually
#torch==2.7.1+cpu
#torchaudio==2.7.1+cpu
#torchinfo==1.8.0
